requirements.txt
googletrans==4.0.0-rc1
discord.py==2.0.0
python-dotenv==1.0.0
yt-dlp==2024.08.06
PyNaCl==1.5.0
openai==0.28.0
langdetect==1.0.9